Refactor CLI to not use yeoman and write tests
Refactor CLI core to not use yeoman. This was done to attempt to isolate
the non-terminated process bug experienced when running migrations. The
issue stems from using inquirer. A workaround has been put in place to
prevent this issue from happening in the future (no hacks); instead of
using the when question setting, questions are removed entirely and the
prompt interface is only called when there are questsions to ask.

Write unit tests for each command line action.